# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocraticAI API
  version: 1.0.0
  description: >
    Backend API contract for SocraticAI (LMS + Socratic tutor).
    This spec is drafted from the PRD and should be treated as the source of truth for frontend integration.
servers:
  - url: http://localhost:4000
tags:
  - name: Health
  - name: Auth
  - name: Courses
  - name: Documents
  - name: Concepts
  - name: Tutor
  - name: Knowledge
  - name: Quiz
  - name: Analytics

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      additionalProperties: false
      required: [success, data, error, meta]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
        meta:
          type: object
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input
    HealthData:
      type: object
      additionalProperties: false
      required: [status, time]
      properties:
        status:
          type: string
          example: ok
        time:
          type: string
          format: date-time
        nodeEnv:
          type: string
          example: development
    SignupRequest:
      type: object
      additionalProperties: false
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        role:
          type: string
          enum: [FACULTY, STUDENT]
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      additionalProperties: false
      required: [token, user]
      properties:
        token: { type: string }
        user:
          type: object
          additionalProperties: false
          required: [id, email, role]
          properties:
            id: { type: string }
            email: { type: string, format: email }
            role: { type: string, enum: [FACULTY, STUDENT] }
    Course:
      type: object
      additionalProperties: false
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        semester: { type: string, nullable: true }
        joinCode: { type: string, nullable: true }
        isActive: { type: boolean }
        facultyId: { type: string }
        createdAt: { type: string, format: date-time }
    CreateCourseRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        semester: { type: string }
    JoinCourseRequest:
      type: object
      additionalProperties: false
      required: [joinCode]
      properties:
        joinCode: { type: string }
    UploadDocumentRequest:
      type: object
      additionalProperties: false
      required: [courseId, file]
      properties:
        courseId: { type: string }
        file:
          type: string
          format: binary
    Document:
      type: object
      additionalProperties: false
      required: [id, courseId, fileName, mimeType, sizeBytes, status, uploadedAt]
      properties:
        id: { type: string }
        courseId: { type: string }
        fileName: { type: string }
        fileUrl: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        status: { type: string, enum: [PENDING, PROCESSING, DONE, FAILED] }
        errorMessage: { type: string, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
        uploadedAt: { type: string, format: date-time }
        chunkCount: { type: integer }
    TutorQuestionRequest:
      type: object
      additionalProperties: false
      required: [courseId, question]
      properties:
        courseId: { type: string }
        question: { type: string }
        sessionId: { type: string, nullable: true }
    TutorQuestionResponse:
      type: object
      additionalProperties: false
      required: [sessionId, mode]
      properties:
        sessionId: { type: string }
        mode:
          type: string
          description: Whether the tutor asks a Socratic prerequisite question or answers directly.
          enum: [SOCRATIC, ANSWER]
        gapConceptId: { type: string, nullable: true }
        message: { type: string }

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check (non-versioned)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/HealthData" }

  /api/v1/health:
    get:
      tags: [Health]
      summary: Health check (versioned)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/HealthData" }

  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Signup
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }

  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [id, email, role]
                        properties:
                          id: { type: string }
                          email: { type: string, format: email }
                          role: { type: string, enum: [FACULTY, STUDENT] }

  /api/v1/courses:
    get:
      tags: [Courses]
      summary: List my courses
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Course" }
    post:
      tags: [Courses]
      summary: Create course (faculty)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourseRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Course" }

  /api/v1/courses/join:
    post:
      tags: [Courses]
      summary: Join course (student)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JoinCourseRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Course" }

  /api/v1/courses/{courseId}:
    get:
      tags: [Courses]
      summary: Get one course if user has access
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Course" }

  /api/v1/documents/upload:
    post:
      tags: [Documents]
      summary: Upload file and queue ingestion (faculty)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/UploadDocumentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Document" }

  /api/v1/documents/course/{courseId}:
    get:
      tags: [Documents]
      summary: List documents and ingestion statuses for a course (faculty)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Document" }

  /api/v1/documents/{id}:
    delete:
      tags: [Documents]
      summary: Delete document and related chunks (faculty)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }

  /api/v1/tutor/question:
    post:
      tags: [Tutor]
      summary: Submit student question (gap detection + Socratic)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TutorQuestionRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/TutorQuestionResponse" }